HS-31-01: decompose db.py into a package + extract MeetingRepository
Split the 5,481-line db.py monolith into the holdspeak/db/ package:
- models.py: the 18 dataclasses + validation constants (shared, kills the
  repo import cycle)
- base.py: BaseRepository (connection factory + shared _json_* helpers)
- meetings.py: MeetingRepository — meetings/segments/speakers/topics/
  bookmarks/action-items + intel snapshots, moved verbatim
- core.py: the MeetingDatabase container (5481 -> 4311), still holding the
  un-migrated domains; constructs self.meetings
- __init__.py: re-exports the full public surface, so `from holdspeak.db
  import X` is unchanged

Greenfield posture: the moved methods are deleted from MeetingDatabase (no
compat facade) and 53 production + 112 test call sites move to db.meetings.*.
One false positive (session.edit_action_item) deliberately left alone; 18
fake-db test doubles get a `meetings` self-property; the duplicate-method
guard test now scans the whole package.

Suite green at 2062 passed / 14 skipped (== baseline); db package ruff-clean.
