+++ holdspeak/tui/services/action_items.py
@@ -11,7 +11,7 @@ def list_action_items(
 ) -> list[ActionItemSummary]:
     """List action items for TUI views."""
     db = db_module.get_database()
-    return db.list_action_items(
+    return db.meetings.list_action_items(
         include_completed=include_completed,
         meeting_id=meeting_id,
         owner=owner,
@@ -21,4 +21,4 @@ def list_action_items(
 def update_action_item_status(action_id: str, status: str) -> bool:
     """Update a persisted action item's status."""
     db = db_module.get_database()
-    return db.update_action_item_status(action_id, status)
+    return db.meetings.update_action_item_status(action_id, status)
