+++ pm/roadmap/holdspeak/README.md
@@ -4,8 +4,8 @@
 > pick up, and the repo conventions that bite (PMO commit gate, write-once
 > evidence, no `Co-Authored-By`, metal-test exclusion).
 
-**Last updated:** 2026-06-02 (Opened **Phase 31 — Database Decomposition** and **Phase 32 — Foundation Hardening & Doc Truth** off a comprehensive engineering review. Phase 31 splits the 5,481-line `db.py` god-object into per-domain repositories behind a thin `Database` container — and, taking the **greenfield/aggressive** posture (one user, one dev DB, destructive-OK), **deletes** the `MeetingDatabase` god-class and **squashes** the 18-version migration ladder to one canonical schema rather than preserving compat. Phase 32 finishes the runtime decomposition Phase 26 left standing (class-ify `web_runtime.py`, invert meeting→web), converges the three audio-ownership models, adds the missing **ungated** CI core-path smoke test, extracts a route error helper, and **deletes** stale **non-PMO** docs (HANDOVER's dead "11 stubs" line; PLAN "branch feature/…" headers for branches that don't exist). The PMO historical record is kept verbatim by design. Suite green at 2075 collected.).
-**Current phase:** **Phase 31** ([phase-31-database-decomposition](./phase-31-database-decomposition/current-phase-status.md)) is the actionable software pickup (HS-31-01 first); **Phase 32** ([phase-32-foundation-hardening](./phase-32-foundation-hardening/current-phase-status.md)) follows it. Still open & **hardware-gated** (author remote): **Phase 25** (HS-25-07 dogfood), **Phase 24** (companion, 3/6), **Phase 15** (out-and-about, gated on Phase 25). Next plugin work is the **frontier** (authoring guide / packs / actuators), not stubs — see Phase 29's `final-summary.md` §Handoff.
+**Last updated:** 2026-06-02 (Opened **Phase 31 — Database Decomposition** and **Phase 32 — Foundation Hardening & Doc Truth** off a comprehensive engineering review. Phase 31 splits the 5,481-line `db.py` god-object into per-domain repositories behind a thin `Database` container — and, taking the **greenfield/aggressive** posture (one user, one dev DB, destructive-OK), **deletes** the `MeetingDatabase` god-class and **squashes** the 18-version migration ladder to one canonical schema rather than preserving compat. Phase 32 finishes the runtime decomposition Phase 26 left standing (class-ify `web_runtime.py`, invert meeting→web), converges the three audio-ownership models, adds the missing **ungated** CI core-path smoke test, extracts a route error helper, and **deletes** stale **non-PMO** docs (HANDOVER's dead "11 stubs" line; PLAN "branch feature/…" headers for branches that don't exist). The PMO historical record is kept verbatim by design. **HS-31-01 shipped same day:** `db.py` → `holdspeak/db/` package (`models`/`base`/`meetings`/`core`), the meetings domain extracted to `MeetingRepository`, 53 production + 112 test call sites moved to `db.meetings.*`, `MeetingDatabase` shed 5481→4311 lines; full public surface re-exported (no caller import changes); suite green at 2062 passed / 14 skipped, db package ruff-clean.).
+**Current phase:** **Phase 31** ([phase-31-database-decomposition](./phase-31-database-decomposition/current-phase-status.md)) — in-progress, **1/5** (HS-31-01 done: `db.py` is now a package, meetings domain extracted to `MeetingRepository`, suite green at 2062; HS-31-02 next). **Phase 32** ([phase-32-foundation-hardening](./phase-32-foundation-hardening/current-phase-status.md)) follows it. Still open & **hardware-gated** (author remote): **Phase 25** (HS-25-07 dogfood), **Phase 24** (companion, 3/6), **Phase 15** (out-and-about, gated on Phase 25). Next plugin work is the **frontier** (authoring guide / packs / actuators), not stubs — see Phase 29's `final-summary.md` §Handoff.
 **Status:** Plugin rollout complete (Phases 16 → 27 → 28 → 29; 14 real plugins). Remaining open work is largely hardware-gated (Phases 24, 25 HS-25-07).
 
 ## Vision
